Remove reference to IE7-specific css and js.
Although the original aim was to try and support IE7 (hence use of
javascript animation rather than css transitions), it has become a
struggle to move the project forward (as a sole developer) due to
dealing with all the snags of IE7. The final straw was the fact the IE7
engine struggles to cope with animating large numbers of elements, so
smoothness is lost. Therefore, I am ditching it.