Expose helpful functions and refactor
Introduce helper functions for programmatic manipulation.
Expose pagination, navigation, presentation subcomponents on self, to
allow functions defined (and yet to be defined) on the prototype access.
Remove handler attachment to individual crumb-outer and frame-outer
items, and instead add a single handler on the respective belt element
and ascertain the target item from the e object, which should reduce
memory footprint.